conf file support
This is implemented in a minimalistic manner. Config files are
expected to be key-value pairs where a value is required and
keys alone where values are not. Values are separated from
keys by '='. Each key value pair must be on a new line.

This also changes flags to use key value pairs, so we can
set local=false in our config, and override with --local true
on the commandline.

Options can be set with both config files and command line
arguments and command line arguments will override options
set in config files. This is achieved by configuring
clap/structopt to allow all params to be overridable and
parsing a config file into a vector of strings in the same
format as command line arguments and prepending that vector
to the command line arguments. For example a conf file with

scheduler-port=1234
local=true

is parsed as ["--scheduler-port", "1234", "--local" "true"]

which is then parsed by structopt. (With the minor detail that
we prepend with the binary name, as structopt expects that)